backend/requirements-lambda.txt
# AWS Lambda Requirements
boto3==1.34.34
botocore==1.34.34

# Amazon Bedrock
anthropic-bedrock==0.8.0

# Image Processing (lightweight for Lambda)
Pillow==10.2.0
numpy==1.24.3

# Medical Image Processing
pydicom==2.4.4
SimpleITK==2.3.1

# Data Processing
pandas==2.0.3

# Utilities
python-dateutil==2.8.2
